Final logic update
finished pristege logic, other than that all is complete

+++ DogeClicker/app/src/main/java/com/example/dogeclicker/models/Event.java
@@ -2,12 +2,12 @@ package com.example.dogeclicker.models;
 
 public class Event {
     private String name;
-    private int effect;
+    private double effect;
     private String promptOfEvent;
     private boolean goodEvent;
     private boolean isSkillPointEffect;
 
-    public Event(String name, int effect, String promptOfEvent, boolean goodEvent,boolean isSkillPointEffect) {
+    public Event(String name, double effect, String promptOfEvent, boolean goodEvent,boolean isSkillPointEffect) {
         setName(name);
         setEffect(effect);
         setPromptOfEvent(promptOfEvent);
@@ -23,11 +23,11 @@ public class Event {
         this.name = name;
     }
 
-    public int getEffect() {
+    public double getEffect() {
         return effect;
     }
 
-    public void setEffect(int effect) {
+    public void setEffect(double effect) {
         this.effect = effect;
     }
